OC-61: resolve code inspector warnings

+++ clover-eclipse/org.openclover.eclipse.core/src/main/java/org/openclover/eclipse/core/projects/model/CoverageFilesUtils.java
@@ -20,8 +20,10 @@ public class CoverageFilesUtils {
                 return false;
             };
             File[] filesToDelete = coverageDbFolder.listFiles(deleteFilter);
-            for (File toDelete : filesToDelete) {
-                toDelete.delete();
+            if (filesToDelete != null) {
+                for (File toDelete : filesToDelete) {
+                    toDelete.delete();
+                }
             }
         }
     }
